# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Quin Course",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cart/addcart/{id}": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add to your cart",
                "description": "",
                "operationId": "addcart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of this course",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/cart/deletecart/{id}": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Delete course from your cart",
                "description": "",
                "operationId": "deletecart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of this course",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/cart/asyn_cart": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Asynchronous cart update",
                "description": "Update courses in your cart asynchronously by providing an array of course IDs.",
                "operationId": "asyn_cart",
                "parameters": [
                    {
                        "name": "id[]",
                        "in": "query",
                        "description": "Array of course IDs",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "number"
                            }
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/cart/my_cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "my_cart ",
                "description": "Get my cart ",
                "operationId": "my_cart",
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/category/get_child/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get category child information by id",
                "description": "Returns category child information",
                "operationId": "getchildcategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id category child",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "JWT Token (Bearer <token>)"
            }
        }
    }
}